chore: Move DiningLocation to a separate file
